[aapt kotlin] flattening/extracting xmls.
This is a change for XmlProcessor to start flattening xml to proto. This
primarily consists of two parts:

First, the XmlProtoBuilder offers a simple way to build flattened xmls
quickly, and also allows easy tracking of active xml namespaces. This is
important as outlined xml "aapt:attr" inherit all the namespaces that
were active in scope of the "aapt:attr" in the original xml file.

i.e.
<View1 xmlns:android="http://schemas.android.com/apk/res/android"
    xmlns:aapt="http://schemas.android.com/aapt">
  <aapt:attr name="android:text">
  <!-- both the "android" and "aapt" namespace should continue -->
  <!-- working in "View2" even when outlined. So they need to -->
  <!-- be declared on the containing element (View2) -->
    <View2 android:text="hey">
      <View3 android:id="hi" />
    </View2>
  </aapt:attr>
</View1>

The second part is all the logic in the XmlProcessor relating to using
the XmlProtoBuilder and outlining Aapt:Attr, which may be nested.

Test: Added tests for both the XmlProtoBuilder and XmlProcessor.
Bug: 132800341
Change-Id: Ib8fb1cc1d7fc67121c2f56e601a930a25b387dc2