Always cache dex merging task outputs
Because task does some non-trivial processing
even when copying input dex files to output, caching
is actually beneficial (it counts the number of files
which incurs FS access and it opens JARs to copy the
entries). This CL removes considitional
caching, and always caches the task outputs.

In order to better understand previous state, here
is some background:

Previous implementation was not doing what it was
meant to do; it was suboptimal. The intention was
to avoid caching when copying inputs to outputs which was:
- when min API < 23 and total number of files is
strictly less than LIBRARIES_MERGING_THRESHOLD
- when min API >= 23 and total number of files is
strictly less than LIBRARIES_M_PLUS_MAX_THRESHOLD

BUG: n/a
Test: existing cacheability tests
Change-Id: I9d44e8cb8709b892aede16459c879d5c53bdaea6