Add incremental run to GenerateLibraryProguardRulesTask
Proguard keep class rules are currently unnecessarily regenerated when a
new layout file is added or a non layout resource is added, changed or
modified. This change stops task runs when non-resources are added,
changed or deleted and incrementaly appends added layout keep
classes. Proguard rules are only regenerated when manifests are
added/modified/removed or when layout files are modified or removed.

Bug: 128593134
Test: GenerateLibraryProguardRulesTaskTest
Change-Id: Ia09e3ed55f027d9a2dc8d757dc914544dd251584