Don't save JUnit results after Windows builds
Starting with 06b9fa85d7e613dbc9fc, we save the test result XML files in
the DIST_DIR. These are copied by buildbot, but that process seems to
take anywhere from 5 to 10 minutes.

This CL reverts that change. We'll re-introduce it once we save the
results in a single zip container.

Bug: 143992474
Test: N/A
Change-Id: Id8363419cd0be3c2d43bc5b15c8521a15fa6c39c