Add base overlay class to installer library
Adds a class to encapsulate operations on an IWI overlay. Adds the
following operations:

- Open() : Opens the overlay for modification, creating it if neccessary
- WriteFile() : Writes a file to the overlay if it is open
- Commit() : Closes the overlay, finalizing any modifications made.

Overlays have an id file, which acts as a hash of the content inside the
overlay. An overlay without an id file is an invalid overlay, and should
be deleted.

To ensure that it is possible to know that modifications to an overlay
have completed successfully, an overlay must be Open()'d before
modification, which deletes the current id file.

This leaves the overlay in an invalid state until a corresponding Commit()
is called, which writes the new id. An overlay without a valid id file
cannot be opened, preventing modifications to broken overlays.

Bug: 144062091
Test: overlay_test.cc
Change-Id: I6b5cc477eeb3b170f53a9bd916cad82e4402ec9a