Add new lint check and quickfix for Kotlin assertions
Android	Lint already warns developers that assertions in Java are not
enforced in Android.

This CL adds a new lint check which adds the same warning for Kotlin
(which uses a runtime library instead of a language facility to
implement assertions).

Furthermore, it implements quickfixes for both the Java and Kotlin
warnings to perform the recommended replacement: replacing
   assert(condition(), { "message" })
with
   if (BuildConfig.DEBUG && !condition()) { error("message") }

Finally, Kotlin assertions differ from Java assertions in that the
arguments (the condition) is always evaluated, whether or not
assertions are off. That means that you should not do any significant
computation to determine the condition, unless you know that
assertions are enabled.

The new Kotlin assertion lint check also enforces this condition (for
when lint is analyzing non-Android code), and offers a quickfix to
surround the assert() call with a check to see if assertions are
enabled:
     assert(expensive())
gets replaced with
     if (javaClass.desiredAssertionStatus()) { assert(expensive()) }

This CL also makes some fixes to the test infrastructure around
quickfixes to better test quickfixes which shortens symbols.

Bug: 145238106
Test: Unit tests included
Change-Id: Ib78a952a9d79f9678a66150f6ac177c8958abaa3