Create Aapt2Workers build service
This CL creates aapt2 workers build service that can
be shared between tasks, and whose lifecycle is managed
by Gradle. Instant execution is incompatible with static
state, which is how this service was stored before.

In addition to that, workers factory is not fully stateless,
which is also necessary for instant execution.

BUG: 145129387
Test: existing
Change-Id: I84755d530dd119b56616dd458afbe925d1bedcda