Add EvictingExecutorService
New ExecutorService that allows having a limited number of tasks
waiting. If new tasks arrive when the maximum number of tasks are
already waiting, the oldest one will be dropped from the queue to make
space for the new one.
This ExecutorService is useful for executing render refresh calls where
new ones superseed the ones waiting in the queue. The EvictingExecutor
requires a delegate ExecutorService to run the tasks.

This CL also adds a new OnDemandExecutorService for testing. This
executor acts as a queue than can be executed on demand.

Change-Id: Ib91981563c4d290e466eb49417657b0f1530c398
Test: Adds tests
Bug: N/A