Redirect classloader resource load for --dont-kill
BaseDexClassLoader / DexPathList internally maintain references to the
APK(s) that were used to load this application. These APKs are used when
calling ClassLoader#getResources(), and are not updated during a
--dont-kill install.

This CL uses instrumentation/reflection to ensure that classloader
resource paths referencing the installed APK are always up-to-date.

The instrumentation only triggers when getResources() is called after a
--dont-kill install has occurred; there should be no impact outside of
this scenario.

Fixes: 130141145
Test: Manual
Test: Get resource from class loader before/after swapping on API 27,
28, 29
Change-Id: Ib04bbad551adf01ba17fed019eef3df81ef76139