Raw File support for kotlin resource compiler
Added support, and enables raw files to be used by the Kotlin Resource
Compiler. The code is fairly straightforward, but here's the jist:

1. build the ResourceFile header from the input.

2. serialize the header

3. write the header and the file contents to flatbuffer.

Also, updated the method that determines which files the Resource
Compiler supports.

Bug: 132800341
Test: ran integration tests with the boolean option enabled and ran the
benchmark for the Jvm resource compiler. Also, updated the
ResourceCompilerTest file to reflect changes and test different effects.
Change-Id: I57d09811e395b6230f40237a2197f05395f074ea