Wrap aapt2 daemon services in Gradle build service
Every task that needs access to aapt2 daemon must obtain
instance of the build service. The build service provides
access to aapt2 daemons and manages their lifecycle.

In order to make daemons accessible from worker actions,
tasks can generate a key that can be used to access worker
action service registry. The only exception to this is
AarResourcesCompilerTransform which accesses the registry
directly (this is because Gradle does not currently support
build services in artifact transforms).

This CL also removes global WorkerActionServiceRegistry.INSTANCE
and local static registries are instantiated as needed.

Bug: 145129387
Test: existing
Change-Id: I3976c4078ec0e85b20f7653da102a8dca4b68d10